[ci] Remove PR-rebased check
This check too frequently results in the develop branch having a
failing CI status even when all the tests are passing, if a PR is
rebased and then quickly merged (or squash-merged via the GitHub web
UI).  The GitHub built-in check for merge conflicts is sufficient to
catch most issues.