Support orderless annotation and kwd matchers. (#17052)
* Support orderless annotation and kwd matchers.

Orderless has a few new matchers that are quite useful. The kwd matcher provides
a lot of different ones based on different modes, and the annotation matcher
with `&` allows you to match against all marginalia annotations.

To enable this we need to change the component separators back to the
default (which is a space not preceded by a backslash).

This could be left as a configuration option, but I figured the nicest solution
was to use the default. People can override if they prefer the other behaviour.

* Keep '&' separator for company completion

(It might be possible to replace the advice by defining a separate completion
style for in-buffer completion, see orderless-define-completion-style and
the discussion in https://github.com/minad/consult/pull/1131.)

---------

Co-authored-by: fnussbaum <ferdinand.nussbaum@inf.ethz.ch>