fix: set next-error-last-buffer when calling pytest
python-pytest does not use `compilation-start` (which would have taken care of
that for us) and neither sets the variable manually.

This kludge works around that by advising python-pytest--get-buffer. This
function is called from a python buffer when running pytest -- so it looks
like the easiest way to intercept the pytest-buffer.