Rework treatment of built-in packages (#16938)
* Move transient to bootstrap layer

Transient is a built-in package, and also a dependency of many other packages
users might install. It makes more sense to always configure it, also when the
git layer is not used.

* [core] Correctly initialize location of built-in packages

Previously, if no location was specified, it was set to the default `elpa`.
This did not mean, however, that the packages where installed from ELPA; the
built-in packages were used regardless (they might not even be published on an
ELPA, of course).

Now we correctly set the location to `built-in`, in preparation to making
`:location elpa` take effect for built-in packages, too.

* [core] Ensure built-in packages with different locations are updated

We need to account for built-in packages when installing, checking availability
of new versions, updating, and rolling back packages.

Also refactor the last call site of configuration-layer//delete-package that
used package names instead of pkg-desc objects (namely
configuration-layer/rollback), such that we can simplify the function.

* [org] Remove `:min-version` for org package

The package will now get updated by `configuration-layer/update-packages`.
We do not need to manually bump the version anymore.

* [core] Prevent outdated built-in packages from being loaded