[core] Suggest fix for obsolete hybrid-mode variables
These variables are unconditionally bound if the
dotspacemacs-editing-style is `hybrid' and the user hasn't updated
their `hybrid-mode' package since the new obsoletion warning was
added.  It is not obvious how to fix this issue because the variables
may not appear in the user's .spacemacs at all, so explicitly suggest
to the user that they should try updating the package.